Refactor API
The API was too specific and was not leading to something that could be
extended. This update removes all the encode functionality and
reimagines the API to be more of a three part tuple of
(name, type, options). This allows it to be extendable by type. More
work needs to be done to use namespaced keys. It will be important
picking the correct namespace so that the API still reads clean when
writing binspecs.