Add compile and decode phases
Originally there was a single decode function that would take a spec and
bits and return the result. This is not as efficient as the spec would
have to be continuously re-conformed. It also was more difficult to
reason about. This update changes this into two phases. The first phase
will call a compile function for each spec entry. A compile function
will take the current spec entry and return a decode function. The
decode funtion takes the current state the current path in this state
and a bitbuffer. This allows decoding to happen much more quickly as the
spec no longer has to parsed.